# Set `Base.Experimental.@max_methods 1`