@rikhuijzer 2022-06-08
> Let's also try to set this in `PlutoRunner` and test how it will go!

I just did and couldn't spot any gain in the number of allocations for the compilation time benchmark. I suspect that it is because there are no cases where max methods is an issue or because the difference is so small that it is hard to spot or it is already done via the global setting for the package. I'm not gonna dive deeper into this